Change all int to long

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
 
-        int[] sales = {1, 5, 71, 13, 0};
+        long[] sales = {1, 5, 71, 13, 0};
 
         SalesManager salesManag = new SalesManager(sales);
